+++ src/lib/server/actions/permohonan-skd.ts
@@ -13,6 +13,7 @@ export async function createPermohonanSKDAction(formData: FormData) {
   try {
     const currentSession = await findCurrentSessionService();
     if (!currentSession?.user) throw new ApiError(status.UNAUTHORIZED, MESSAGE.GLOBAL.UNAUTHORIZED);
+    if (currentSession.user.role !== "ADMIN") throw new ApiError(status.FORBIDDEN, "Hanya admin yang dapat melakukan aksi ini");
 
     const rawData = Object.fromEntries(formData.entries());
     const parsedData = {
@@ -35,6 +36,7 @@ export async function updatePermohonanSKDAction(permohonanSKDId: string, formDat
   try {
     const currentSession = await findCurrentSessionService();
     if (!currentSession?.user) throw new ApiError(status.UNAUTHORIZED, MESSAGE.GLOBAL.UNAUTHORIZED);
+    if (currentSession.user.role !== "ADMIN") throw new ApiError(status.FORBIDDEN, "Hanya admin yang dapat melakukan aksi ini");
 
     const rawData = Object.fromEntries(formData.entries());
     const parsedData = {
@@ -57,6 +59,7 @@ export async function deletePermohonanSKDAction(permohonanSKDId: string) {
   try {
     const currentSession = await findCurrentSessionService();
     if (!currentSession?.user) throw new ApiError(status.UNAUTHORIZED, MESSAGE.GLOBAL.UNAUTHORIZED);
+    if (currentSession.user.role !== "ADMIN") throw new ApiError(status.FORBIDDEN, "Hanya admin yang dapat melakukan aksi ini");
 
     await deletePermohonanSKDService(permohonanSKDId);
     revalidatePath(PATHS.SKD_REQUEST);
@@ -71,6 +74,7 @@ export async function followUpPermohonanSKDAction(formData: FormData) {
   try {
     const currentSession = await findCurrentSessionService();
     if (!currentSession?.user) throw new ApiError(status.UNAUTHORIZED, MESSAGE.GLOBAL.UNAUTHORIZED);
+    if (currentSession.user.role !== "ADMIN") throw new ApiError(status.FORBIDDEN, "Hanya admin yang dapat melakukan aksi ini");
 
     const rawData = Object.fromEntries(formData.entries());
     const validatedData = followUpPermohonanSKDSchema.parse(rawData);
