+++ src/lib/server/actions/permohonan-skk.ts
@@ -22,6 +22,7 @@ export const createPermohonanSKKAction = async (
   try {
     const currentSession = await findCurrentSessionService();
     if (!currentSession?.user) throw new ApiError(status.UNAUTHORIZED, MESSAGE.AUTH.UNAUTHORIZED);
+    if (currentSession.user.role !== "ADMIN") throw new ApiError(status.FORBIDDEN, "Hanya admin yang dapat melakukan aksi ini");
     const data = await createPermohonanSKKService(currentSession.user.userId, payload);
     revalidatePath(PATHS.SKK_REQUEST);
     return { status: status.OK, message: message.CREATE_OK, data };
@@ -38,6 +39,7 @@ export const updatePermohonanSKKAction = async (
   try {
     const currentSession = await findCurrentSessionService();
     if (!currentSession?.user) throw new ApiError(status.UNAUTHORIZED, MESSAGE.AUTH.UNAUTHORIZED);
+    if (currentSession.user.role !== "ADMIN") throw new ApiError(status.FORBIDDEN, "Hanya admin yang dapat melakukan aksi ini");
     const data = await updatePermohonanSKKService(id, payload);
     revalidatePath(PATHS.SKK_REQUEST);
     return { status: status.OK, message: message.UPDATE_OK, data };
@@ -51,6 +53,7 @@ export const followUpPermohonanSKKAction = async (values: unknown) => {
   try {
     const currentSession = await findCurrentSessionService();
     if (!currentSession?.user) throw new ApiError(status.UNAUTHORIZED, MESSAGE.AUTH.UNAUTHORIZED);
+    if (currentSession.user.role !== "ADMIN") throw new ApiError(status.FORBIDDEN, "Hanya admin yang dapat melakukan aksi ini");
 
     const { permohonanSKKId, ...validatedData } = followUpPermohonanSKKSchema.parse(values);
     const data = await followUpPermohonanSKKService(permohonanSKKId, validatedData);
@@ -66,6 +69,7 @@ export const deletePermohonanSKKAction = async (id: string) => {
   try {
     const currentSession = await findCurrentSessionService();
     if (!currentSession?.user) throw new ApiError(status.UNAUTHORIZED, MESSAGE.AUTH.UNAUTHORIZED);
+    if (currentSession.user.role !== "ADMIN") throw new ApiError(status.FORBIDDEN, "Hanya admin yang dapat melakukan aksi ini");
     const data = await deletePermohonanSKKService(id);
     revalidatePath(PATHS.SKK_REQUEST);
     return { status: status.OK, message: message.DELETE_OK, data };
